refactor: consolidate business logic in FmcdCore and fix balance calculation
  - Fix zero balance bug by using client.get_balance() instead of only counting mint notes
  - Remove duplicate implementations between FmcdCore and API layers
  - Implement PaymentInfoResolver trait pattern for LNURL handling without coupling core to web protocols
  - Move all business logic (get_info, join_federation, create_invoice, pay_invoice) to FmcdCore
  - Make API layer a thin wrapper that delegates to FmcdCore methods
  - Add proper error handling and context propagation throughout
