+++ src/api/resolvers/mod.rs
@@ -1,0 +1,3 @@
+pub mod payment;
+
+pub use payment::LnurlResolver;
