+++ src/api/rest/ln/mod.rs
@@ -1,71 +1,17 @@
-use std::str::FromStr;
-
-use anyhow::{bail, Context};
+use anyhow::bail;
 use fedimint_client::ClientHandleArc;
 use fedimint_core::Amount;
 use fedimint_ln_client::{InternalPayState, LightningClientModule, LnPayState, PayType};
-use fedimint_ln_common::lightning_invoice::Bolt11Invoice;
 use futures_util::StreamExt;
 use tracing::{debug, info};
 
-use self::pay::{LnPayRequest, LnPayResponse};
-use crate::observability::sanitize_invoice;
+use crate::core::LnPayResponse;
 
 pub mod gateways;
 pub mod invoice;
 pub mod pay;
 pub mod status;
 pub mod stream;
-
-pub async fn get_invoice(req: &LnPayRequest) -> anyhow::Result<Bolt11Invoice> {
-    let info = req.payment_info.trim();
-    match Bolt11Invoice::from_str(info) {
-        Ok(invoice) => {
-            debug!(
-                "Parsed parameter as bolt11 invoice: {}",
-                sanitize_invoice(&invoice)
-            );
-            match (invoice.amount_milli_satoshis(), req.amount_msat) {
-                (Some(_), Some(_)) => {
-                    bail!("Amount specified in both invoice and command line")
-                }
-                (None, _) => {
-                    bail!("We don't support invoices without an amount")
-                }
-                _ => {}
-            };
-            Ok(invoice)
-        }
-        Err(e) => {
-            let lnurl = if info.to_lowercase().starts_with("lnurl") {
-                lnurl::lnurl::LnUrl::from_str(info)?
-            } else if info.contains('@') {
-                lnurl::lightning_address::LightningAddress::from_str(info)?.lnurl()
-            } else {
-                bail!("Invalid invoice or lnurl: {e:?}");
-            };
-            debug!("Parsed parameter as lnurl: {lnurl:?}");
-            let amount = req
-                .amount_msat
-                .context("When using a lnurl, an amount must be specified")?;
-            let async_client = lnurl::AsyncClient::from_client(reqwest::Client::new());
-            let response = async_client.make_request(&lnurl.url).await?;
-            match response {
-                lnurl::LnUrlResponse::LnUrlPayResponse(response) => {
-                    let invoice = async_client
-                        .get_invoice(&response, amount.msats, None, req.lnurl_comment.as_deref())
-                        .await?;
-                    let invoice = Bolt11Invoice::from_str(invoice.invoice())?;
-                    assert_eq!(invoice.amount_milli_satoshis(), Some(amount.msats));
-                    Ok(invoice)
-                }
-                other => {
-                    bail!("Unexpected response from lnurl: {other:?}");
-                }
-            }
-        }
-    }
-}
 
 pub async fn wait_for_ln_payment(
     client: &ClientHandleArc,
