+++ src/core/operations/mod.rs
@@ -4,6 +4,3 @@ pub use payment::{InvoiceTracker, PaymentState, PaymentTracker};
 
 #[cfg(test)]
 mod tests;
-
-// Re-export the payment module contents for easier access
-pub use payment::*;
